Add a more sophisticated tunnel generation algorithm
It is still not perfect, but it will be more than enough for now.
However, the entries will cluster in the middle of the virtual circle in
which they get generated. Only testing inside the game will tell if
it is too obvious that they belong to the same tunnel.