Add mice spawner
At the start, all mice get randomly spawned in one tunnel except the
win tunnel. At the moment they spawn at one exit but in the future it
would be more realistic to spawn them between two exits.

Also the mice do not move and since they are spawned hidden inside
tunnels they will not be sent to the client.