Add collision detection and mouse catching
Cats can now catch mice and mice get properly cached. Mice only move
straight down and don't even try to reach the goal. The reason for
that is, that this will be already quite handy for developing the client
to show mice and developing a sophisticated navigation algorithm will be
a lot easier with a client that can display mice.