Drastically improve path finding performance by removing dicts
By removing the costs dict and emplacing it directly into the node I
was able to cut time from ~300ms to ~10ms (tested with 30 mice and
~60 exits)

That is quite a crazy performance improvement and while the runtime
hasn't changed algorithmicly, it makes sense that dereferncing a field
inside a struct is quite faster than hashing and searching in a
hashtable.

We still aren't in the time we want to hit, but with path caching it
should be quite possible. ^^