Introduce path caching
This isn't as big of a improvement as hoped, and only brings a ~2x
speed improvement. Currently, this brings us into the the time budget
for 30 mice with ~60 exits.

I expect the tick time to get worse with cat avoidance, but since the
server get compiled with optimization, I don't think we are going to
keep having a lot problems.