Make cat order send deterministic
The server now tracks when each user joined and sends all updates to
the client with the cats in the order they joined.

This required a larger change in the protocol, because in the
gamescoreboard we were using a HashMap which is unordered (shout-out to
`std::unordered_map`). Therefore I had to replace it with an array and a
new intermediate class.